chore: Update the style of the legends + cleanup
