+++ web/src/components/MapOverlays.tsx
@@ -3,16 +3,14 @@ import { useAtomValue } from 'jotai';
 import { lazy, Suspense } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { hasSeenSurveyCardAtom } from 'utils/state/atoms';
+import { useIsBiggerThanMobile } from 'utils/styling';
 
 import SurveyCard from './app-survey/SurveyCard';
 import LegendContainer from './legend/LegendContainer';
 
-function FallbackComponent() {
-  return <div />;
-}
-
 export default function MapOverlays() {
   const hasSeenSurveyCard = useAtomValue(hasSeenSurveyCardAtom);
+  const isBiggerThanMobile = useIsBiggerThanMobile();
 
   const [searchParameters] = useSearchParams();
   const showManager =
@@ -25,18 +23,22 @@ export default function MapOverlays() {
 
   const FeatureFlagsManager = showManager
     ? lazy(() => import('features/feature-flags/FeatureFlagsManager'))
-    : FallbackComponent;
+    : () => undefined;
 
   return (
-    <div className="pointer-events-none  fixed top-12 z-20  m-3 flex flex-col items-end space-y-3 sm:bottom-0 sm:right-0 sm:top-auto ">
-      <Suspense>
-        <div className="hidden sm:flex">{showManager && <FeatureFlagsManager />}</div>
-      </Suspense>
-      <Suspense>
-        <div className="hidden sm:flex">
-          <LegendContainer />
-        </div>
-      </Suspense>
+    <div className="pointer-events-none fixed top-12 z-20 m-3 flex flex-col items-end space-y-3 sm:bottom-0 sm:right-0 sm:top-auto ">
+      {isBiggerThanMobile && (
+        <>
+          {showManager && (
+            <Suspense>
+              <FeatureFlagsManager />
+            </Suspense>
+          )}
+          <Suspense>
+            <LegendContainer />
+          </Suspense>
+        </>
+      )}
       {surveyEnabled && (
         <Suspense>
           <SurveyCard />
